+++ src/main/java/io/github/devatherock/artifactory/service/DockerBadgeService.java
@@ -6,6 +6,7 @@ import io.github.devatherock.artifactory.entities.ArtifactoryFolderElement;
 import io.github.devatherock.artifactory.entities.ArtifactoryFolderInfo;
 import io.github.devatherock.artifactory.entities.DockerLayer;
 import io.github.devatherock.artifactory.entities.DockerManifest;
+import io.github.devatherock.artifactory.util.BadgeGenerator;
 import io.micronaut.cache.annotation.Cacheable;
 import io.micronaut.core.util.CollectionUtils;
 import io.micronaut.http.HttpMethod;
@@ -26,15 +27,15 @@ public class DockerBadgeService {
     private static final String HDR_API_KEY = "X-JFrog-Art-Api";
 
     private final BlockingHttpClient artifactoryClient;
-    private final ShieldsIOClient shieldsIOClient;
+    private final BadgeGenerator badgeGenerator;
     private final ArtifactoryProperties artifactoryConfig;
 
     @Inject
     public DockerBadgeService(@Client("${artifactory.url}") HttpClient artifactoryClient,
-                              ShieldsIOClient shieldsIOClient,
+                              BadgeGenerator badgeGenerator,
                               ArtifactoryProperties artifactoryConfig) {
         this.artifactoryClient = artifactoryClient.toBlocking();
-        this.shieldsIOClient = shieldsIOClient;
+        this.badgeGenerator = badgeGenerator;
         this.artifactoryConfig = artifactoryConfig;
     }
 
@@ -50,7 +51,7 @@ public class DockerBadgeService {
                     .get() / BYTES_IN_MB;
 
             LOGGER.info("Size of {}/{}: {} MB", packageName, tag, imageSize);
-            return shieldsIOClient.downloadBadge(String.format("%s-%.2f MB-blue.svg", badgeLabel, imageSize));
+            return badgeGenerator.generateBadge(badgeLabel, String.format("%.2f MB", imageSize));
         } else {
             return generateNotFoundBadge(badgeLabel);
         }
@@ -62,8 +63,7 @@ public class DockerBadgeService {
 
         if (null != manifest && CollectionUtils.isNotEmpty(manifest.getLayers())) {
             LOGGER.info("Layers in {}/{}: {}", packageName, tag, manifest.getLayers().size());
-            return shieldsIOClient.downloadBadge(String.format("%s-%d-blue.svg", badgeLabel,
-                    manifest.getLayers().size()));
+            return badgeGenerator.generateBadge(badgeLabel, String.valueOf(manifest.getLayers().size()));
         } else {
             return generateNotFoundBadge(badgeLabel);
         }
@@ -92,15 +92,14 @@ public class DockerBadgeService {
                 }
             }
             LOGGER.info("Download count of {}: {}", packageName, downloadCount);
-            return shieldsIOClient.downloadBadge(String.format("%s-%d-blue.svg", badgeLabel,
-                    downloadCount));
+            return badgeGenerator.generateBadge(badgeLabel, String.valueOf(downloadCount));
         } else {
             return generateNotFoundBadge(badgeLabel);
         }
     }
 
     private String generateNotFoundBadge(String badgeLabel) {
-        return shieldsIOClient.downloadBadge(String.format("%s-Not Found-blue.svg", badgeLabel));
+        return badgeGenerator.generateBadge(badgeLabel, "Not Found");
     }
 
     private DockerManifest readManifest(String packageName, String tag) {